+++ cpython/src/gen/java/org/bytedeco/cpython/PyGetSetDef.java
@@ -24,9 +24,9 @@ public class PyGetSetDef extends Pointer {
         return (PyGetSetDef)super.position(position);
     }
 
-    @MemberGetter public native @Cast("const char*") BytePointer name();
+    public native @Cast("const char*") BytePointer name(); public native PyGetSetDef name(BytePointer setter);
     public native getter get(); public native PyGetSetDef get(getter setter);
     public native setter set(); public native PyGetSetDef set(setter setter);
-    @MemberGetter public native @Cast("const char*") BytePointer doc();
+    public native @Cast("const char*") BytePointer doc(); public native PyGetSetDef doc(BytePointer setter);
     public native Pointer closure(); public native PyGetSetDef closure(Pointer setter);
 }